Move HTMLWriter.RD into a separate file
